lab-jenkins: init script
